requirements.txt
alembic==1.16.5
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
certifi==2025.8.3
celery==5.3.6
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.8.0
email-validator==2.3.0
flake8==7.3.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
pytz
pytz==2024.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.14.0
pyflakes==3.4.0
pyotp==2.9.0
python-owasp-zap-v2.4==0.1.0
requests==2.32.4
redis==5.0.1
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.42
typing_extensions==4.14.1
urllib3==2.5.0
weasyprint
Werkzeug==3.1.3
WTForms==3.2.1
zaproxy==0.4.0
